netdog: modularize wicked dhcp code
This commit begins the process of modularizing the `wicked` code to
prepare for future additions.  The structure will largely follow the
structure in `net_config` to make the code easy to correlate.  Currently
only the DHCP-related wicked code is split out; more to come.
